% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgetoadj.R
\name{ed2ad}
\alias{ed2ad}
\title{Convert \link[qgraph]{qgraph} edge list to adjacency matrix}
\usage{
ed2ad(netob)
}
\arguments{
\item{netob}{A qgraph object.}
}
\value{
A matrix. Returns the adjacency matrix.
}
\description{
Convenience function to convert \link[qgraph]{qgraph} edge list to adjacency matrix.
}
\examples{
\dontrun{ 
data(Obama)
CorMat <- qgraph::cor_auto(Obama[,1:25])
EBICgraph <- qgraph::qgraph(CorMat, graph = "glasso", sampleSize = nrow(bfi),
                tuning = 0.5, layout = "spring", title = "BIC", details = TRUE)
adjmat <- ed2ad(EBICgraph)}

}
\author{
Pedro Henrique Ribeiro Santiago, \email{pedro.ribeirosantiago@adelaide.edu.au} [ctb]

Gustavo Hermes Soares, \email{gustavo.soares@adelaide.edu.au} [rev]

Adrian Quintero, \email{adrianquintero987@hotmail.com} [rev]

Lisa Jamieson, \email{lisa.jamieson@adelaide.edu.au} [rev]
}
