% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printcpAlgorithm.R
\name{print.cpAlgorithm}
\alias{print.cpAlgorithm}
\title{print.cpAlgorithm}
\usage{
\method{print}{cpAlgorithm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cpAlgorithm}; see also
\link{cpAlgorithm}}

\item{...}{currently ignored}
}
\description{
Print method for objects of class \code{cpAlgorithm}.
}
\author{
Jens Lange, \email{lange.jens@outlook.com}
}
