% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printcpPermuteEntropy.R
\name{print.cpPermuteEntropy}
\alias{print.cpPermuteEntropy}
\title{print.cpPermuteEntropy}
\usage{
\method{print}{cpPermuteEntropy}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cpPermuteEntropy}; see also
\link{cpPermuteEntropy}}

\item{...}{currently ignored}
}
\description{
Print method for objects of class \code{cpPermuteEntropy}.
}
\author{
Jens Lange, \email{lange.jens@outlook.com}
}
